+++ insights-ui/src/scripts/industry-tariff-reports/05-industry-areas.ts
@@ -1,7 +1,7 @@
+import { writeJsonFileForIndustryAreaSections, writeMarkdownFileForIndustryAreaSections } from '@/scripts/industry-tariff-reports/tariff-report-read-write';
+import { IndustryAreaSection, IndustryAreasWrapper } from '@/scripts/industry-tariff-reports/tariff-types';
 import { getLlmResponse, outputInstructions } from '@/scripts/llm-utils';
-import { IndustryAreasWrapper, IndustryAreaSection } from '@/scripts/industry-tariff-reports/tariff-types';
 import { z } from 'zod';
-import { uploadFileToS3, getJsonFromS3 } from '@/scripts/report-file-utils';
 
 const IndustryAreaSectionSchema = z.object({
   title: z.string().describe('Title of the section which discusses various industry areas.'),
@@ -40,35 +40,9 @@ async function getIndustryAreaSection(industry: string, headings: IndustryAreasW
   return response;
 }
 
-function getS3Key(industry: string, fileName: string): string {
-  return `koalagains-reports/tariff-reports/${industry.toLowerCase()}/05-industry-areas/${fileName}`;
-}
-
 export async function getAndWriteIndustryAreaSectionToJsonFile(industry: string, headings: IndustryAreasWrapper): Promise<void> {
   const industryAreaSection = await getIndustryAreaSection(industry, headings);
 
-  // Upload JSON to S3
-  const jsonKey = getS3Key(industry, 'industry-area.json');
-  await uploadFileToS3(new TextEncoder().encode(JSON.stringify(industryAreaSection, null, 2)), jsonKey, 'application/json');
-
-  // Generate and upload markdown
-  const markdownContent = getMarkdownContentForIndustryAreas(industryAreaSection);
-  const markdownKey = getS3Key(industry, 'industry-area.md');
-  await uploadFileToS3(new TextEncoder().encode(markdownContent), markdownKey, 'text/markdown');
-}
-
-export async function readIndustryAreaSectionFromFile(industry: string): Promise<IndustryAreaSection | undefined> {
-  const key = getS3Key(industry, 'industry-area.json');
-  return await getJsonFromS3<IndustryAreaSection>(key);
-}
-
-export function getMarkdownContentForIndustryAreas(industryAreaSection: IndustryAreaSection) {
-  const markdownContent = `# ${industryAreaSection.title}\n\n${industryAreaSection.industryAreas}`;
-  return markdownContent;
-}
-
-export async function writeIndustryAreaSectionToMarkdownFile(industry: string, industryAreaSection: IndustryAreaSection): Promise<void> {
-  const markdownContent = getMarkdownContentForIndustryAreas(industryAreaSection);
-  const key = getS3Key(industry, 'industry-area.md');
-  await uploadFileToS3(new TextEncoder().encode(markdownContent), key, 'text/markdown');
+  await writeJsonFileForIndustryAreaSections(industry, industryAreaSection);
+  await writeMarkdownFileForIndustryAreaSections(industry, industryAreaSection);
 }
